# perf(db): Improve speed of query in `clear_expired_resolutions`