This query is slow because it does an inefficient join to `sentry_release_project`. Just fetching
the ids separately and doing an in seems to improve things.

This produces a query plan like
```
QUERY PLAN
Nested Loop  (cost=22.74..1141.39 rows=1 width=44) (actual time=21.057..21.057 rows=0 loops=1)
  Join Filter: (sentry_groupresolution.release_id = sentry_release.id)
  ->  Nested Loop  (cost=22.18..1125.39 rows=5 width=52) (actual time=21.057..21.057 rows=0 loops=1)
        ->  Bitmap Heap Scan on sentry_groupresolution  (cost=21.62..436.37 rows=265 width=44) (actual time=2.476..8.305 rows=5121 loops=1)
              Recheck Cond: (((status = 0) AND (type = 1)) OR ((status = 0) AND (type IS NULL)))
              Filter: (release_id <> 81950264)
              Heap Blocks: exact=1478
              ->  BitmapOr  (cost=21.62..21.62 rows=265 width=0) (actual time=2.272..2.272 rows=0 loops=1)
                    ->  Bitmap Index Scan on idx_sentry_groupresolution_pending  (cost=0.00..19.06 rows=264 width=0) (actual time=2.263..2.263 rows=5131 loops=1)
                          Index Cond: ((status = 0) AND (type = 1))
                    ->  Bitmap Index Scan on idx_sentry_groupresolution_pending  (cost=0.00..2.43 rows=1 width=0) (actual time=0.008..0.008 rows=3 loops=1)
                          Index Cond: ((status = 0) AND (type IS NULL))
        ->  Index Only Scan using sentry_release_project_project_id_35add08b8e678ec7_uniq on sentry_release_project  (cost=0.56..2.59 rows=1 width=8) (actual time=0.002..0.002 rows=0 loops=5121)
              Index Cond: ((project_id = 1519906) AND (release_id = sentry_groupresolution.release_id))
              Heap Fetches: 0
  ->  Index Scan using sentry_release_id_organization_id on sentry_release  (cost=0.56..3.19 rows=1 width=4) (never executed)
        Index Cond: (id = sentry_release_project.release_id)
        Filter: ((date_added < '2019-10-05 00:02:04.369+00'::timestamp with time zone) AND (date_added > '2019-09-05 00:02:04.369+00'::timestamp with time zone))
Planning time: 2.697 ms
Execution time: 21.167 ms
```

Instead of 
```
QUERY PLAN
Nested Loop  (cost=134.67..668.09 rows=1 width=44) (actual time=1243.231..1243.231 rows=0 loops=1)
  Join Filter: (sentry_groupresolution.release_id = sentry_release.id)
  ->  Hash Join  (cost=134.11..667.46 rows=1 width=52) (actual time=1243.229..1243.229 rows=0 loops=1)
        Hash Cond: (sentry_groupresolution.release_id = sentry_release_project.release_id)
        ->  Bitmap Heap Scan on sentry_groupresolution  (cost=21.62..436.37 rows=265 width=44) (actual time=2.441..7.842 rows=5123 loops=1)
              Recheck Cond: (((status = 0) AND (type = 1)) OR ((status = 0) AND (type IS NULL)))
              Filter: (release_id <> 81950264)
              Heap Blocks: exact=1478
              ->  BitmapOr  (cost=21.62..21.62 rows=265 width=0) (actual time=2.207..2.207 rows=0 loops=1)
                    ->  Bitmap Index Scan on idx_sentry_groupresolution_pending  (cost=0.00..19.06 rows=264 width=0) (actual time=2.198..2.198 rows=5133 loops=1)
                          Index Cond: ((status = 0) AND (type = 1))
                    ->  Bitmap Index Scan on idx_sentry_groupresolution_pending  (cost=0.00..2.43 rows=1 width=0) (actual time=0.008..0.008 rows=3 loops=1)
                          Index Cond: ((status = 0) AND (type IS NULL))
        ->  Hash  (cost=80.90..80.90 rows=2527 width=8) (actual time=1234.600..1234.600 rows=187364 loops=1)
              Buckets: 262144 (originally 4096)  Batches: 1 (originally 1)  Memory Usage: 9367kB
              ->  Nested Loop  (cost=10.29..80.90 rows=2527 width=8) (actual time=0.235..1196.363 rows=187364 loops=1)
                    ->  Unique  (cost=9.73..9.74 rows=2 width=16) (actual time=0.134..0.136 rows=1 loops=1)
                          ->  Sort  (cost=9.73..9.73 rows=2 width=16) (actual time=0.133..0.134 rows=1 loops=1)
                                Sort Key: u0.id
                                Sort Method: quicksort  Memory: 25kB
                                ->  Nested Loop  (cost=0.86..9.72 rows=2 width=16) (actual time=0.101..0.102 rows=1 loops=1)
                                      ->  Index Scan using sentry_release_project_release_id on sentry_release_project u1  (cost=0.44..4.81 rows=2 width=8) (actual time=0.038..0.038 rows=1 loops=1)
                                            Index Cond: (release_id = 81950264)
                                      ->  Index Only Scan using sentry_project_pkey on sentry_project u0  (cost=0.42..2.44 rows=1 width=8) (actual time=0.058..0.059 rows=1 loops=1)
                                            Index Cond: (id = u1.project_id)
                                            Heap Fetches: 1
                    ->  Index Only Scan using sentry_release_project_project_id_35add08b8e678ec7_uniq on sentry_release_project  (cost=0.56..22.94 rows=1264 width=16) (actual time=0.098..1174.575 rows=187364 loops=1)
                          Index Cond: (project_id = u0.id)
                          Heap Fetches: 158755
  ->  Index Scan using sentry_release_id_organization_id on sentry_release  (cost=0.56..0.62 rows=1 width=4) (never executed)
        Index Cond: (id = sentry_release_project.release_id)
        Filter: (date_added < '2019-10-05 00:02:04.369+00'::timestamp with time zone)
Planning time: 3.473 ms
Execution time: 1243.555 ms
```